requirements.txt
PyQt6>=6.5.0
requests>=2.28.0
aiohttp>=3.8.0
Pillow>=9.5.0
python-dotenv>=1.0.0
keyboard>=0.13.5
pynput>=1.7.6
setuptools>=65.5.0
python-i18n>=0.3.9
pyperclip>=1.8.2
openai>=1.0.0 